pyproject.toml
[tool.poetry]
name = "generate-nfts"
version = "0.1.0"
description = ""
authors = ["NajeebShafi <najeeb,shafi@yahoo.com>"]
readme = "README.md"
packages = [{include = "generate_nfts"}]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.23.3"
pandas = "^1.5.0"
Pillow = "^9.2.0"
progressbar = "^2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

